Moved major version code to shared project
