+++ .gitignore
@@ -85,3 +85,5 @@ src/scaffolding.config
 # Approval tests temp file
 *.received.*
 .connectionstrings.config
+
+**/TestRun*
