--- src/NHPersistenceCompatibilityTests/4.5/TestSagaDataWithComposite.cs
@@ -1,18 +1,0 @@
-﻿using System;
-using NServiceBus.Saga;
-
-namespace Version_4_5
-{
-    public class TestSagaDataWithComposite : IContainSagaData
-    {
-        public virtual Guid Id { get; set; }
-        public virtual string Originator { get; set; }
-        public virtual string OriginalMessageId { get; set; }
-        public virtual SagaComposite Composite { get; set; }
-
-        public class SagaComposite
-        {
-            public virtual string Value { get; set; }
-        }
-    }
-}
