--- src/NHPersistenceCompatibilityTests/6.2/NHibernateSessionFactory.cs
@@ -1,35 +1,0 @@
-﻿using System;
-using Common;
-using NHibernate;
-using NHibernate.Tool.hbm2ddl;
-using NServiceBus.SagaPersisters.NHibernate.AutoPersistence;
-
-namespace Version_6_2
-{
-    public class NHibernateSessionFactory
-    {
-        public NHibernateSessionFactory()
-        {
-            SessionFactory = new Lazy<ISessionFactory>(Init);
-        }
-
-        private ISessionFactory Init()
-        {
-            var configuration = new NHibernate.Cfg.Configuration().AddProperties(NHibernateConnectionInfo.Settings);
-            var modelMapper = new SagaModelMapper(new[]
-            {
-                typeof(TestSagaDataWithList),
-                typeof(TestSagaDataWithComposite),
-                typeof(TestSagaData)
-            });
-
-            configuration.AddMapping(modelMapper.Compile());
-
-            new SchemaUpdate(configuration).Execute(false, true);
-
-            return configuration.BuildSessionFactory();
-        }
-
-        public Lazy<ISessionFactory> SessionFactory { get; }
-    }
-}
