--- src/NHPersistenceCompatibilityTests/6.2/TestSagaData.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using NServiceBus.Saga;
-
-namespace Version_6_2
-{
-    public class TestSagaData : IContainSagaData
-    {
-        public virtual Guid Id { get; set; }
-        public virtual string Originator { get; set; }
-        public virtual string OriginalMessageId { get; set; }
-    }
-}
