--- src/NHPersistenceCompatibilityTests/7.0/TestSagaData.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using NServiceBus;
-
-namespace Version_7_0
-{
-    public class TestSagaData : IContainSagaData
-    {
-        public virtual Guid Id { get; set; }
-        public virtual string Originator { get; set; }
-        public virtual string OriginalMessageId { get; set; }
-    }
-}
