+++ src/NHPersistenceCompatibilityTests/PersistenceTests.sln
@@ -1,7 +1,7 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 14
-VisualStudioVersion = 14.0.24720.0
+VisualStudioVersion = 14.0.25123.0
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "7.0", "7.0\7.0.csproj", "{59859466-3F15-40A0-9A9B-003096CCAD02}"
 EndProject
@@ -15,7 +15,16 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "5.0", "5.0\5.0.csproj", "{0
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "4.5", "4.5\4.5.csproj", "{FF4E993B-A2DA-4BEB-9DE2-33D33F208543}"
 EndProject
+Project("{D954291E-2A0B-460D-934E-DC6B0785DB48}") = "Shared", "Shared\Shared.shproj", "{5E568185-F3A6-4546-B63D-E35BE391F349}"
+EndProject
 Global
+	GlobalSection(SharedMSBuildProjectFiles) = preSolution
+		Shared\Shared.projitems*{096b1d0d-2d8b-4cb5-a26d-8377da1812ce}*SharedItemsImports = 4
+		Shared\Shared.projitems*{67e400df-1fa1-4b04-a985-68f187c03610}*SharedItemsImports = 4
+		Shared\Shared.projitems*{5e568185-f3a6-4546-b63d-e35be391f349}*SharedItemsImports = 13
+		Shared\Shared.projitems*{59859466-3f15-40a0-9a9b-003096ccad02}*SharedItemsImports = 4
+		Shared\Shared.projitems*{ff4e993b-a2da-4beb-9de2-33d33f208543}*SharedItemsImports = 4
+	EndGlobalSection
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
 		Release|Any CPU = Release|Any CPU
