+++ src/NHPersistenceCompatibilityTests/Shared/NHibernateSessionFactory.cs
@@ -3,9 +3,11 @@ using Common;
 using NHibernate;
 using NHibernate.Tool.hbm2ddl;
 using NServiceBus.SagaPersisters.NHibernate.AutoPersistence;
+#if NHVersion7
 using NServiceBus.Sagas;
+#endif
 
-namespace Version_7_0
+namespace Shared
 {
     public class NHibernateSessionFactory
     {
@@ -17,7 +19,7 @@ namespace Version_7_0
         private ISessionFactory Init()
         {
             var configuration = new NHibernate.Cfg.Configuration().AddProperties(NHibernateConnectionInfo.Settings);
-
+#if NHVersion7
             var metaModel = new SagaMetadataCollection();
             metaModel.Initialize(new[] { typeof(TestSaga) });
             var metadata = metaModel.Find(typeof(TestSaga));
@@ -35,7 +37,16 @@ namespace Version_7_0
             metadata = metaModel.Find(typeof(TestSagaWithComposite));
             mapper = new SagaModelMapper(metaModel, new[] { metadata.SagaEntityType });
             configuration.AddMapping(mapper.Compile());
+#else
+            var modelMapper = new SagaModelMapper(new[]
+            {
+                typeof(TestSagaData),
+                typeof(TestSagaDataWithComposite),
+                typeof(TestSagaDataWithList)
+            });
 
+            configuration.AddMapping(modelMapper.Compile());
+#endif
             new SchemaUpdate(configuration).Execute(false, true);
 
             return configuration.BuildSessionFactory();
