+++ src/NHPersistenceCompatibilityTests/Shared/Shared.projitems
@@ -1,0 +1,18 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<Project xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
+  <PropertyGroup>
+    <MSBuildAllProjects>$(MSBuildAllProjects);$(MSBuildThisFileFullPath)</MSBuildAllProjects>
+    <HasSharedItems>true</HasSharedItems>
+    <SharedGUID>5e568185-f3a6-4546-b63d-e35be391f349</SharedGUID>
+  </PropertyGroup>
+  <PropertyGroup Label="Configuration">
+    <Import_RootNamespace>Shared</Import_RootNamespace>
+  </PropertyGroup>
+  <ItemGroup>
+    <Compile Include="$(MSBuildThisFileDirectory)NHibernateSessionFactory.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)TestSaga.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)TestSagaData.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)TestSagaDataWithComposite.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)TestSagaDataWithList.cs" />
+  </ItemGroup>
+</Project>
