+++ src/NHPersistenceCompatibilityTests/Shared/TestSagaData.cs
@@ -1,14 +1,16 @@
 ﻿using System;
-using System.Collections.Generic;
+#if NHVersion7
+using NServiceBus;
+#else
 using NServiceBus.Saga;
+#endif
 
-namespace Version_6_2
+namespace Shared
 {
-    class TestSagaDataWithList : IContainSagaData
+    public class TestSagaData : IContainSagaData
     {
         public virtual Guid Id { get; set; }
         public virtual string Originator { get; set; }
         public virtual string OriginalMessageId { get; set; }
-        public virtual IList<int> Ints { get; set; }
     }
 }
